Bugfix: Generate melody to match entire chord progression
For each chord progression, melodies were being generated until the
chord progression (4 measures) were filled with notes. The problem is,
it was starting at the first chord and generating notes until the 4
measures filled up without moving onto the other chords, so when the
other chords were playing, it was actually playing melodies written for
the first chord, making it sound bad. This fix still allows notes to
overflow into other chords (e.g. we could have a dotted whole note) but
it will move on and start generating the melody for the proper chord
after that overflow note has passed.